serde_multipart

Function serialize_bincode

pub fn serialize_bincode<S: ?Sized + Serialize>(
    value: &S,
) -> Result<Message, Error>

Serialize the provided value into a multipart message. The value is encoded using an extended version of bincode that skips serializing Parts, which are instead held directly by the returned message.

Serialize uses the same codec options as bincode::serialize / bincode::deserialize. These are currently not customizable unless an explicit specialization is also provided.